@@ -2,19 +2,29 @@ name: zoom_web
 description: Web platform implementation of flutter_zoom.
 version: 1.0.2+1
 repository: https://github.com/driftboat/flutter_zoom/tree/main/zoom_web
+publish_to: "none"
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: ">=2.17.0 <3.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  zoom_platform_interface: ^1.0.2
+  # zoom_platform_interface: ^1.0.2
+  zoom_platform_interface:
+    git:
+      url: https://github.com/Gaurav192/flutter_zoom
+      ref: sdkKey_web
+      path: zoom_platform_interface
   js: ^0.6.3
-  
+
+dependency_overrides:
+  zoom_platform_interface:
+    path: "../zoom_platform_interface"
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -30,11 +40,11 @@ flutter:
   # adding or updating assets for this project.
   plugin:
     platforms:
-    # This plugin project was generated without specifying any
-    # platforms with the `--platform` argument. If you see the `some_platform` map below, remove it and
-    # then add platforms following the instruction here:
-    # https://flutter.dev/docs/development/packages-and-plugins/developing-packages#plugin-platforms
-    # -------------------
+      # This plugin project was generated without specifying any
+      # platforms with the `--platform` argument. If you see the `some_platform` map below, remove it and
+      # then add platforms following the instruction here:
+      # https://flutter.dev/docs/development/packages-and-plugins/developing-packages#plugin-platforms
+      # -------------------
       web:
         pluginClass: ZoomWeb
         fileName: zoom_web.dart
